fixed logging of blank files

+++ moveLogs.sh
@@ -1,6 +1,16 @@
 #!/bin/bash
+if [ -s log.txt ]
+then
+	cp log.txt "logs/log$(date +"%m-%d-%y").txt"
+	> log.txt
+else
+	echo "logs empty"
+fi
 
-cp log.txt "logs/log$(date +"%m-%d-%y").txt"
-> log.txt
-cp nohup.out "logs/nohup$(date +"%m-%d-%y").out"
-> nohup.out
+if [ -s nohup.out ]
+then
+	cp nohup.out "logs/nohup$(date +"%m-%d-%y").out"
+	> nohup.out
+else
+	echo "nohup empty"
+fi
